backend/requirements.txt
fastapi[standard]>=0.111.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
